cql3: internal select statements
Running execute() will try to talk to endpoints and other structures, that may
not be initialized by the time we execute internal queries - for the simple fact
that we issue internal queries when setting up most of that data.

Implement execute_internal instead, which will keep everything node-local, without
ever trying to go towards other nodes.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>