column_family: add the commitlog as a parameter
When we create a column family, we can pass as an extra (optional) parameter,
the commitlog. Because the commitlog is optional to begin with - it won't exist
if we don't call init_commitlog, we can have this to be empty meaning no commit log.

The creation of a column family should be always done through
add_column_family. And if that is the case, we have the database's commitlog
right there and can get the pointer through the db. Only tests are not creating
the column family this way, and for them, it is fine.

We want to do that, because some column family operations will use the commit log.
Right now, they are forcing us to add parameters to APIs that would be much cleaner
without it. So while separation is good, this level of coupling is a net win as it
allows us to clean up some visible APIs.

Another alternative to this would be to just pass a db pointer. From that db
pointer, we could grab the commit log or everything else we may need in the
future. I will argue against it, though, since passing the full database is a
bit of an overkill, and would force us to use a reference attribute or naked
pointer - because we certainly can't just move the db into a shared_ptr. More
importantly, the fact that this change is not visible, but kept internally in
the database method, allow us to cleanly modify it if we really need to in the
future.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>