types: get_hashed_name
Sometimes, Origin will store collections in strings like this:

    636f6c756d6e:org.apache.cassandra.db.marshal.ListType(org.apache.cassandra.db.marshal.Int32Type)

The number you see before the colon is the hash used internally to store
specific instances of that type.

This patch provides a method that allows us to query for such string
when needed. The basic name() method stays the same. It is not always
that we want the string to be in this format.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>