cell_name: add basic types
As much as the types themselves have names, Origin builds on top of that a "CellName".
The cell name encodes information about the types used to parse the clustering keys,
and are usually things like:

    org.apache.cassandra.db.marshal.CompositeType(types...),

for a composite type, for instance.

This is not importing code from Origin, because I am mostly adapting to what we have.
But the code is largely inspired on it.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>