type_parser: parse CompositeTypes
Some type strings, like comparator strings, contains more than just the type
information.  In particular, the string that describes a comparator, will
indicate whether or not it is a composite by the usage of CompositeType, and
will wrap its collections around ColumnToCollectionType.

This technically doesn't have to be part of the type_parser, but by making it
so, we are able to reuse a lot of its inner machinery.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>