sstables: simplified version of write_column_name for non-clustered columns
We still want to wrap it instead of writing the column name directly, so we are
able to update the statistics.

It is better to have a separate function for this, because write_column_name
doesn't have enough information to decide when to do what. Augmenting it so we
could have would require passing the schema, or an extra parameter, which would
then spread to all callers.

Keep in mind that testing for an empty clustering key is not enough, since
composite types will serialize the empty clustering key in this case.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>