compound: invert order for byte comparison of reversed types
reversed types can be byte_comparable, but in this case we should
invert the order of the comparation.

One alternative here, of course, would be to just declare all reversed types
non-byte comparable. That would definitely be safer, but at the expense of
always having more expensive comparisons for inverted orders.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>