sstable_test: test clustering order
If we revert the type of the clustering key, which is what would happen if we
defined the table as with clustering order by (cl desc), we expect the
clustering keys to be in descending order on disk.

There is no work needed for sstables for that to happen. But we should still
verify that this is indeed the case.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>