db: implement control groups
Right now, there is exactly one control group per shard, which serves
to throttle requests when the amount of dirty memory is above a threshold.

This is rather coarse, as the database is non-responsive while we wait
for memory to be cleaned, but it's more responsive than a crashed database.
Future updates will take into account the sawtooth behavior of memory
allocation and freeing.