create directory for commitlog
The same way we create a directory for the sstables, we should do it for the
commitlog.

If you, like me, have been using the --datadir directive, the fact that we
don't do it would have slipped away: when using --datadir, both the commitlog
and sstables will be put in this same root.

However, when using the system defaults from the configuration file, this won't
be true.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>