utils: provide access to the rpc address
Doing what we do here is not ideal. It suffers from the same problems as the
broadcast address, and it is actually a little bit worse: since there is no
natural place for us to fetch the rpc address from, within the network classes,
we have to record it once we set it.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>