sstable_test: fix check_toc_func
We are currently failing the sstable test. The reason is that we use the store()
function for test purposes, and that function does not store the TOC component.
It was removed by Aviccident in 3a5e3c88.

Because that function is only used for testing purposes, it doesn't need to write
the Index and Data components: we can then remove them from the list.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>