row consumer: extend usage of the read_* functions
In some places, we cannot use our read_* functions, because we don't know
whether or not it succeeded, and that is important when passing the state
along.

The fix for this is trivial, since we can just return it from the reader.

Note for reviewers: The commend in one of the functions say we should use:
"read_bytes(data, _u32, _key ...". But in the actual code, the where buffer is
_val, not _key.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>