row consumer: extend use of read for multi-value fields
In what I can only suppose to be an attempt to gain some cycles, we
are testing whether we can read many values at once, and if so, using
consume_be directly for those.

This, unfortunately, steps in the way of the decoupling work I am pursuing.
While the simplest solution would be to just read a single value in here and
set the state to the next, relying on the next loop to process correctly, I can
only assume Nadav had his reasons to code it the way it is.

What we can do then, is read the first value, set the intermediate state
unconditionally, and then, if the first read didn't wait, try the second in the
same loop, and so on, if needed.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>