compaction_manager: fix compaction_manager::stop
For stopping a task of compaction manager, we first close the gate
used by compaction then bust semaphore via semaphore::broken().

The problem is that semaphore::broken() only signals waiters, and so
subsequent semaphore::wait() calls would succeed and the task would
remain alive forever.
The fix is to signal semaphore, forcing the task to exit via gate
exception, if it wasn't waiting on the semaphore by the time that
semaphore::broken() was called by compaction_manager::task_stop().

Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>