storage_service: Kill seastar thread usage in join_token_ring
join_token_ring is called in two places, one is
storage_service::init_server and the other is
storage_service::join_ring.

The former is already inside a seastar thread. The latter is not but it
is rarely called. We can make join_ring runs inside a seastar thread, so
that join_token_ring always runs inside a seastar thread and we can get
rid of creating a thread inside join_token_ring.