filter tracker: move map reduce operations to sstables
The sstable will give us access to the filter tracker, but in a slightly
confusing way: It holds the distributed object, but the distributed object
public methods themselves issue map reduce operations.

Accessing them through the sstables would require an acessor for the local
filter tracker - or to each statistic separately, that then access the result
of the map reduce operation itself.

However, that would make things extremely confusing if we ever want to access
local operations from the sstable as well.

It is a much cleaner interface to move the map reduce operations to the
sstables.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>