api/column family: extend map_reduce_cf to accept futurized Mappers
the simple for loop won't cut for the case the Mapper function returns a future.
The best way to accumulate results for futures (another MR), OTOH, is not a match
for MR.

While we can obviously convert the return value of Mapper to a future, so we
can MR it, we need to at least distinguish between the situation in which a
future - which may or may not be ready is passed, and the case we need to
create one.

And if we are going to go through the trouble, better have a simpler function
for the future case anyway.

So, this function uses SFINAE to allow the compiler to expose to variants of
the all-classes map_reduce_cf: one in which mapper is a future, and one in
which it isn't.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>