api/column family: sstable_stat, an accessor for sstable statistics
This is a helper function to aid us in the task of collecting statistics from
sstables, that refer to the whole column family.

To do that, we always need to iterate over all sstables, and accumulate
accordingly.  All the common code is factored out, and then only in the last
minute, through the call operator, we decide whether to call the named or
unnamed version of map_reduce_cf.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>