api/column family: off-heap usage information
We don't really support the concept of off-heap allocation, at least not in the
same way that the JVM does. While we do have different between allocators -
managed vs non-managed memory, and possibly others, I am not sure we should be
overloading the original java-based statistics with this - specially since we
don't actually intend to expose those as tunable choices.

I will be happy to change that after some discussion if you guys disagree and
reason about it satisfactorily.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>