commonize the NSM
In order to reuse the NSM in other scenarios, we need to push as much code
as possible into a common class.

This patch does that, making the continuous_data_consumer class now the main
placeholder for the NSM class. The actual readers will have to inherit from it.

However, despite using inheritance, I am not using virtual functions (at least
not for the main processor): instead, the function is passed as a parameter for
the constructor, and can then just be called in the processing loop.

Using a virtual function here would have been simpler, but it would also have
cost us 1.77 % performance in the sequential read microbenchmark. Passing the
function as a parameter costs us nothing - the new results are still within the
two error margins.

verify_end_state is not called often. So that one is made virtual.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>