sstables: be more tolerant of errors in the Statistics file
It's great to have statistics, but assert is too big of a hammer. We don't need
to crash due to the lack of it, and can try our best to continue.

We currently have a problem (described in 265), in which we, for some reason,
fail to read the Statistics file. Yes, that is bad: but we can totally go on
without it.

This of course is not the real fix. The bug still exist. But if we fix this,
there can be others.  We can be a bit more resilient overall.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>