Introduce sstable_deletion_manager
sstable deletion is tricky: some sstables are exclusive to this shard
(those that were created by the shard), while others are shared among
all shards.  Exclusive sstables can be deleted as soon as it is determined
that they are unneeded, while shared sstables must be permitted for deletion
by all shards.

The sstable deletion manager solves this problem.  It provides a
synchronization point for deleting shared sstables, and for tracking
deletions in progress.

Two variants are provided: a normal one, and a dummy for unit tests.