column_family: incremental backups
Only tables that arise from flushes are backed up. Compacted tables are not.
Therefore, the place for that to happen is right after our flush.

Note that due to our sharded architecture, it is possible that in the face of a
value change some shards will backup sstables while others won't.

This is, in theory, possible to mitigate through a rwlock. However, this
doesn't differ from the situation where all tables are coming from a single
shard and the toggle happens in the middle of them.

The code as is guarantees that we'll never partially backup a single sstable,
so that is enough of a guarantee.

Signed-off-by: Glauber Costa <glommer@scylladb.com>