snapshots: update information about snapshot existence
We are keeping track of snapshot existence at the keyspace level. When we
create a new snapshot, we should update that. The snapshot() cf method needs to
be updated to take a reference to the database so it can do it itself:
requiring the callers to update it themselves is just waiting for one of them
to forget it.

Signed-off-by: Glauber Costa <glommer@scylladb.com>