storage_service: delete a snapshot
This patch provides an storage service api to delete an snapshot.  Following
our previous work on this, we will assume that all keyspaces and CFs are
visible in all shards. This will allow us to fetch the list of keyspaces in the
present shard and issue the filesystem operations in that same shard.

That simplifies the code tremendously, and because there are not any operations
we need to do previous to the fs ones (like in the case of create snapshot), we
need no synchronization. Even easier.

After the FS operation succeeds, however, we need to modify the state of all shards
notifying them that this snapshot is no more.

Signed-off-by: Glauber Costa <glommer@scylladb.com>