database: get_snapshot_details
For each of the snapshots available, the api may query for some information:
the total size on disk, and the "real" size. As far as I could understand, the
real size is the size that is used by the SSTables themselves, while the total
size includes also the metadata about the snapshot - like the manifest.json
file.

The code in C* just scans through the directory and recursively adds the size
of every file it encounters. We will do the same, but more focused on the
snapshot's needs.

Signed-off-by: Glauber Costa <glommer@scylladb.com>