snapshot manager
This patch introduces the snapshot manager. It allows us to keep a list of all
the snapshots currently tracked by the system.  This will allow us to easily
figure out whether a snapshot exists or not in a certain keyspace. Note that
the only thing we track about the snapshots is their names.

The alternative to using it is going to the filesystem, which involves some
fairly expensive operations in comparison to just (maybe) going to another CPU
and consulting a set. We only keep the snapshot's name in the set, and they
tend to be <= 16 bytes strings. This means that for 1K snapshots, we're still
talking ~16K memory consumed by this, which is a reasonable tradeoff.

Signed-off-by: Glauber Costa <glommer@scylladb.com>